requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.7
dateparser==1.1.0
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
python-binance==1.0.15
python-dateutil==2.8.2
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.26.0
six==1.16.0
tornado==6.1
typing-extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
ujson==4.2.0
urllib3==1.26.7
websockets==9.1
yarl==1.7.2
